+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/acckeeper/ResetController.java
@@ -1,0 +1,56 @@
+package com.dwarfeng.familyhelper.webapi.node.controller.v1.acckeeper;
+
+import com.dwarfeng.familyhelper.webapi.stack.service.acckeeper.ResetResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
+import com.dwarfeng.subgrade.sdk.interceptor.permission.PermissionRequired;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * 重置控制器。
+ *
+ * @author DwArFeng
+ * @since 2.0.0
+ */
+@RestController("acckeeperResetController")
+@RequestMapping("/api/v1/acckeeper")
+public class ResetController {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(ResetController.class);
+
+    private final ResetResponseService service;
+
+    private final ServiceExceptionMapper sem;
+
+    public ResetController(
+            @Qualifier("acckeeperResetResponseServiceImpl") ResetResponseService service,
+            ServiceExceptionMapper sem
+    ) {
+        this.service = service;
+        this.sem = sem;
+    }
+
+    @PostMapping("/reset-protect")
+    @BehaviorAnalyse
+    @LoginRequired
+    @PermissionRequired("webapi.controller_permitted.acckeeper.reset.reset_protect")
+    public FastJsonResponseData<Object> resetProtect(HttpServletRequest request) {
+        try {
+            service.resetProtect();
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            LOGGER.warn("Controller 异常, 信息如下: ", e);
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+}